+++ creme/metrics/recall.py
@@ -1,4 +1,5 @@
 import collections
+import functools
 import statistics
 
 from .. import stats
@@ -7,10 +8,28 @@ from . import base
 from . import precision
 
 
-__all__ = ['Recall', 'MacroRecall', 'MicroRecall']
+__all__ = [
+    'MacroRecall',
+    'MicroRecall',
+    'Recall',
+    'RollingMacroRecall',
+    'RollingMicroRecall',
+    'RollingRecall'
+]
 
 
-class Recall(stats.Mean, base.BinaryClassificationMetric):
+class BaseRecall:
+
+    @property
+    def bigger_is_better(self):
+        return True
+
+    @property
+    def requires_labels(self):
+        return True
+
+
+class Recall(stats.Mean, BaseRecall, base.BinaryClassificationMetric):
     """Binary recall score.
 
     Example:
@@ -33,13 +52,35 @@ class Recall(stats.Mean, base.BinaryClassificationMetric):
 
     """
 
-    @property
-    def bigger_is_better(self):
-        return True
+    def update(self, y_true, y_pred):
+        if y_true:
+            return super().update(y_true == y_pred)
+        return self
 
-    @property
-    def requires_labels(self):
-        return True
+
+class RollingRecall(stats.RollingMean, BaseRecall, base.BinaryClassificationMetric):
+    """Rolling binary recall score.
+
+    Example:
+
+        ::
+
+            >>> from creme import metrics
+            >>> from sklearn.metrics import recall_score
+
+            >>> y_true = [True, False, True, True, True]
+            >>> y_pred = [True, True, False, True, True]
+
+            >>> metric = metrics.RollingRecall(window_size=3)
+            >>> for y_t, y_p in zip(y_true, y_pred):
+            ...     print(metric.update(y_t, y_p).get())
+            1.0
+            1.0
+            0.5
+            0.666666...
+            0.666666...
+
+    """
 
     def update(self, y_true, y_pred):
         if y_true:
@@ -47,7 +88,7 @@ class Recall(stats.Mean, base.BinaryClassificationMetric):
         return self
 
 
-class MacroRecall(base.MultiClassificationMetric):
+class MacroRecall(BaseRecall, base.MultiClassificationMetric):
     """Macro-average recall score.
 
     Example:
@@ -79,14 +120,6 @@ class MacroRecall(base.MultiClassificationMetric):
         self.recalls = collections.defaultdict(Recall)
         self.classes = set()
 
-    @property
-    def bigger_is_better(self):
-        return True
-
-    @property
-    def requires_labels(self):
-        return True
-
     def update(self, y_true, y_pred):
         self.recalls[y_true].update(True, y_true == y_pred)
         self.classes.update({y_true, y_pred})
@@ -97,6 +130,38 @@ class MacroRecall(base.MultiClassificationMetric):
             0 if c not in self.recalls else self.recalls[c].get()
             for c in self.classes
         ))
+
+
+class RollingMacroRecall(MacroRecall):
+    """Rolling macro-average recall score.
+
+    Example:
+
+        ::
+
+            >>> from creme import metrics
+            >>> from sklearn.metrics import recall_score
+
+            >>> y_true = [0, 1, 2, 2, 2]
+            >>> y_pred = [0, 0, 2, 2, 1]
+
+            >>> metric = metrics.RollingMacroRecall(window_size=3)
+            >>> for y_t, y_p in zip(y_true, y_pred):
+            ...     print(metric.update(y_t, y_p).get())
+            1.0
+            0.5
+            0.6666666666666666
+            0.6666666666666666
+            0.5555555555555556
+
+            >>> metric
+            RollingMacroRecall: 0.555556
+
+    """
+
+    def __init__(self, window_size):
+        self.recalls = collections.defaultdict(functools.partial(RollingRecall, window_size))
+        self.classes = set()
 
 
 class MicroRecall(precision.MicroPrecision):
@@ -130,6 +195,40 @@ class MicroRecall(precision.MicroPrecision):
 
     References:
 
+        1. `Why are precision, recall and F1 score equal when using micro averaging in a multi-class problem? <https://simonhessner.de/why-are-precision-recall-and-f1-score-equal-when-using-micro-averaging-in-a-multi-class-problem/>`_
+
+    """
+
+
+class RollingMicroRecall(precision.RollingMicroPrecision):
+    """Rolling micro-average recall score.
+
+    The micro-average recall is exactly equivalent to the micro-average precision as well as the
+    micro-average F1 score.
+
+    Example:
+
+        ::
+
+            >>> from creme import metrics
+            >>> from sklearn.metrics import recall_score
+
+            >>> y_true = [0, 1, 2, 2, 2]
+            >>> y_pred = [0, 0, 2, 2, 1]
+
+            >>> metric = metrics.RollingMicroRecall(window_size=3)
+            >>> for y_t, y_p in zip(y_true, y_pred):
+            ...     print(metric.update(y_t, y_p).get())
+            1.0
+            0.5
+            0.666666...
+            0.666666...
+            0.666666...
+
+            >>> metric
+            RollingMicroRecall: 0.666667
+
+    References:
         1. `Why are precision, recall and F1 score equal when using micro averaging in a multi-class problem? <https://simonhessner.de/why-are-precision-recall-and-f1-score-equal-when-using-micro-averaging-in-a-multi-class-problem/>`_
 
     """