+++ creme/metrics/rmse.py
@@ -1,7 +1,7 @@
 from . import mse
 
 
-__all__ = ['RMSE']
+__all__ = ['RMSE', 'RollingRMSE']
 
 
 class RMSE(mse.MSE):
@@ -26,6 +26,32 @@ class RMSE(mse.MSE):
 
             >>> metric
             RMSE: 0.612372
+
+    """
+
+    def get(self):
+        return super().get() ** 0.5
+
+
+class RollingRMSE(mse.RollingMSE):
+    """Rolling root mean squared error.
+
+    Example:
+
+        ::
+
+            >>> from creme import metrics
+
+            >>> y_true = [3, -0.5, 2, 7]
+            >>> y_pred = [2.5, 0.0, 2, 8]
+
+            >>> metric = metrics.RollingRMSE(window_size=2)
+            >>> for y_t, y_p in zip(y_true, y_pred):
+            ...     print(metric.update(y_t, y_p))
+            RollingRMSE: 0.5
+            RollingRMSE: 0.5
+            RollingRMSE: 0.353553
+            RollingRMSE: 0.707107
 
     """